Created RemoteUploadeService and performed the code changes
