+++ server/src/main/java/org/opensearch/index/shard/RemoteStoreRefreshListener.java
@@ -93,12 +93,14 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
     private volatile Iterator<TimeValue> backoffDelayIterator;
     private final SegmentReplicationCheckpointPublisher checkpointPublisher;
     private final RemoteStoreSettings remoteStoreSettings;
+    private final RemoteUploaderService remoteUploaderService;
 
     public RemoteStoreRefreshListener(
         IndexShard indexShard,
         SegmentReplicationCheckpointPublisher checkpointPublisher,
         RemoteSegmentTransferTracker segmentTracker,
-        RemoteStoreSettings remoteStoreSettings
+        RemoteStoreSettings remoteStoreSettings,
+        RemoteUploaderService remoteUploaderService
     ) {
         super(indexShard.getThreadPool());
         logger = Loggers.getLogger(getClass(), indexShard.shardId());
@@ -106,6 +108,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         this.storeDirectory = indexShard.store().directory();
         this.remoteDirectory = (RemoteSegmentStoreDirectory) ((FilterDirectory) ((FilterDirectory) indexShard.remoteStore().directory())
             .getDelegate()).getDelegate();
+        this.remoteUploaderService = remoteUploaderService;
         localSegmentChecksumMap = new HashMap<>();
         RemoteSegmentMetadata remoteSegmentMetadata = null;
         if (indexShard.routingEntry().primary()) {
@@ -117,7 +120,8 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         }
         // initializing primary term with the primary term of latest metadata in remote store.
         // if no metadata is present, this value will be initialized with -1.
-        this.primaryTerm = remoteSegmentMetadata != null ? remoteSegmentMetadata.getPrimaryTerm() : INVALID_PRIMARY_TERM;
+        this.primaryTerm = remoteSegmentMetadata != null ? remoteSegmentMetadata.getPrimaryTerm() :
+            INVALID_PRIMARY_TERM;
         this.segmentTracker = segmentTracker;
         resetBackOffDelayIterator();
         this.checkpointPublisher = checkpointPublisher;
@@ -156,7 +160,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
     protected boolean performAfterRefreshWithPermit(boolean didRefresh) {
         boolean successful;
         if (shouldSync(didRefresh, false)) {
-            successful = syncSegments();
+            successful = remoteUploaderService.syncSegments();
         } else {
             successful = true;
         }
@@ -212,157 +216,6 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         return false;
     }
 
-    /*
-     @return false if retry is needed
-     */
-    private boolean syncSegments() {
-        if (isReadyForUpload() == false) {
-            // Following check is required to enable retry and make sure that we do not lose this refresh event
-            // When primary shard is restored from remote store, the recovery happens first followed by changing
-            // primaryMode to true. Due to this, the refresh that is triggered post replay of translog will not go through
-            // if following condition does not exist. The segments created as part of translog replay will not be present
-            // in the remote store.
-            return indexShard.state() != IndexShardState.STARTED || !(indexShard.getEngine() instanceof InternalEngine);
-        }
-        beforeSegmentsSync();
-        long refreshTimeMs = segmentTracker.getLocalRefreshTimeMs(), refreshClockTimeMs = segmentTracker.getLocalRefreshClockTimeMs();
-        long refreshSeqNo = segmentTracker.getLocalRefreshSeqNo();
-        long bytesBeforeUpload = segmentTracker.getUploadBytesSucceeded(), startTimeInNS = System.nanoTime();
-        final AtomicBoolean successful = new AtomicBoolean(false);
-
-        try {
-            try {
-                initializeRemoteDirectoryOnTermUpdate();
-                // if a new segments_N file is present in local that is not uploaded to remote store yet, it
-                // is considered as a first refresh post commit. A cleanup of stale commit files is triggered.
-                // This is done to avoid delete post each refresh.
-                if (isRefreshAfterCommit()) {
-                    remoteDirectory.deleteStaleSegmentsAsync(indexShard.getRemoteStoreSettings().getMinRemoteSegmentMetadataFiles());
-                }
-
-                try (GatedCloseable<SegmentInfos> segmentInfosGatedCloseable = indexShard.getSegmentInfosSnapshot()) {
-                    SegmentInfos segmentInfos = segmentInfosGatedCloseable.get();
-                    final ReplicationCheckpoint checkpoint = indexShard.computeReplicationCheckpoint(segmentInfos);
-                    if (checkpoint.getPrimaryTerm() != indexShard.getOperationPrimaryTerm()) {
-                        throw new IllegalStateException(
-                            String.format(
-                                Locale.ROOT,
-                                "primaryTerm mismatch during segments upload to remote store [%s] != [%s]",
-                                checkpoint.getPrimaryTerm(),
-                                indexShard.getOperationPrimaryTerm()
-                            )
-                        );
-                    }
-                    // Capture replication checkpoint before uploading the segments as upload can take some time and checkpoint can
-                    // move.
-                    long lastRefreshedCheckpoint = ((InternalEngine) indexShard.getEngine()).lastRefreshedCheckpoint();
-                    Collection<String> localSegmentsPostRefresh = segmentInfos.files(true);
-
-                    // Create a map of file name to size and update the refresh segment tracker
-                    Map<String, Long> localSegmentsSizeMap = updateLocalSizeMapAndTracker(localSegmentsPostRefresh).entrySet()
-                        .stream()
-                        .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-                    CountDownLatch latch = new CountDownLatch(1);
-                    ActionListener<Void> segmentUploadsCompletedListener = new LatchedActionListener<>(new ActionListener<>() {
-                        @Override
-                        public void onResponse(Void unused) {
-                            try {
-                                logger.debug("New segments upload successful");
-                                // Start metadata file upload
-                                uploadMetadata(localSegmentsPostRefresh, segmentInfos, checkpoint);
-                                logger.debug("Metadata upload successful");
-                                clearStaleFilesFromLocalSegmentChecksumMap(localSegmentsPostRefresh);
-                                onSuccessfulSegmentsSync(
-                                    refreshTimeMs,
-                                    refreshClockTimeMs,
-                                    refreshSeqNo,
-                                    lastRefreshedCheckpoint,
-                                    localSegmentsSizeMap,
-                                    checkpoint
-                                );
-                                // At this point since we have uploaded new segments, segment infos and segment metadata file,
-                                // along with marking minSeqNoToKeep, upload has succeeded completely.
-                                successful.set(true);
-                            } catch (Exception e) {
-                                // We don't want to fail refresh if upload of new segments fails. The missed segments will be re-tried
-                                // as part of exponential back-off retry logic. This should not affect durability of the indexed data
-                                // with remote trans-log integration.
-                                logger.warn("Exception in post new segment upload actions", e);
-                            }
-                        }
-
-                        @Override
-                        public void onFailure(Exception e) {
-                            logger.warn("Exception while uploading new segments to the remote segment store", e);
-                        }
-                    }, latch);
-
-                    // Start the segments files upload
-                    uploadNewSegments(localSegmentsPostRefresh, localSegmentsSizeMap, segmentUploadsCompletedListener);
-                    if (latch.await(
-                        remoteStoreSettings.getClusterRemoteSegmentTransferTimeout().millis(),
-                        TimeUnit.MILLISECONDS
-                    ) == false) {
-                        throw new SegmentUploadFailedException("Timeout while waiting for remote segment transfer to complete");
-                    }
-                } catch (EngineException e) {
-                    logger.warn("Exception while reading SegmentInfosSnapshot", e);
-                }
-            } catch (IOException e) {
-                // We don't want to fail refresh if upload of new segments fails. The missed segments will be re-tried
-                // as part of exponential back-off retry logic. This should not affect durability of the indexed data
-                // with remote trans-log integration.
-                logger.warn("Exception while uploading new segments to the remote segment store", e);
-            }
-        } catch (Throwable t) {
-            logger.error("Exception in RemoteStoreRefreshListener.afterRefresh()", t);
-        }
-        updateFinalStatusInSegmentTracker(successful.get(), bytesBeforeUpload, startTimeInNS);
-        // If there are failures in uploading segments, then we should retry as search idle can lead to
-        // refresh not occurring until write happens.
-        logger.debug("syncSegments runStatus={}", successful.get());
-        return successful.get();
-    }
-
-    /**
-     * Clears the stale files from the latest local segment checksum map.
-     *
-     * @param localSegmentsPostRefresh list of segment files present post refresh
-     */
-    private void clearStaleFilesFromLocalSegmentChecksumMap(Collection<String> localSegmentsPostRefresh) {
-        localSegmentChecksumMap.keySet()
-            .stream()
-            .filter(file -> !localSegmentsPostRefresh.contains(file))
-            .collect(Collectors.toSet())
-            .forEach(localSegmentChecksumMap::remove);
-    }
-
-    private void beforeSegmentsSync() {
-        // Start tracking total uploads started
-        segmentTracker.incrementTotalUploadsStarted();
-    }
-
-    private void onSuccessfulSegmentsSync(
-        long refreshTimeMs,
-        long refreshClockTimeMs,
-        long refreshSeqNo,
-        long lastRefreshedCheckpoint,
-        Map<String, Long> localFileSizeMap,
-        ReplicationCheckpoint checkpoint
-    ) {
-        // Update latest uploaded segment files name in segment tracker
-        segmentTracker.setLatestUploadedFiles(localFileSizeMap.keySet());
-        // Update the remote refresh time and refresh seq no
-        updateRemoteRefreshTimeAndSeqNo(refreshTimeMs, refreshClockTimeMs, refreshSeqNo);
-        // Reset the backoffDelayIterator for the future failures
-        resetBackOffDelayIterator();
-        // Set the minimum sequence number for keeping translog
-        indexShard.getEngine().translogManager().setMinSeqNoToKeep(lastRefreshedCheckpoint + 1);
-        // Publishing the new checkpoint which is used for remote store + segrep indexes
-        checkpointPublisher.publish(indexShard, checkpoint);
-        logger.debug("onSuccessfulSegmentsSync lastRefreshedCheckpoint={} checkpoint={}", lastRefreshedCheckpoint, checkpoint);
-    }
-
     /**
      * Resets the backoff delay iterator so that the next set of failures starts with the base delay and goes upto max delay.
      */
@@ -380,6 +233,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         return ThreadPool.Names.REMOTE_REFRESH_RETRY;
     }
 
+    //todo: create a common class
     private boolean isRefreshAfterCommit() throws IOException {
         String lastCommittedLocalSegmentFileName = SegmentInfos.getLastCommitSegmentsFileName(storeDirectory);
         return (lastCommittedLocalSegmentFileName != null
@@ -399,70 +253,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         return false;
     }
 
-    void uploadMetadata(Collection<String> localSegmentsPostRefresh, SegmentInfos segmentInfos, ReplicationCheckpoint replicationCheckpoint)
-        throws IOException {
-        final long maxSeqNo = ((InternalEngine) indexShard.getEngine()).currentOngoingRefreshCheckpoint();
-        SegmentInfos segmentInfosSnapshot = segmentInfos.clone();
-        Map<String, String> userData = segmentInfosSnapshot.getUserData();
-        userData.put(LOCAL_CHECKPOINT_KEY, String.valueOf(maxSeqNo));
-        userData.put(SequenceNumbers.MAX_SEQ_NO, Long.toString(maxSeqNo));
-        segmentInfosSnapshot.setUserData(userData, false);
-
-        Translog.TranslogGeneration translogGeneration = indexShard.getEngine().translogManager().getTranslogGeneration();
-        if (translogGeneration == null) {
-            throw new UnsupportedOperationException("Encountered null TranslogGeneration while uploading metadata to remote segment store");
-        } else {
-            long translogFileGeneration = translogGeneration.translogFileGeneration;
-            remoteDirectory.uploadMetadata(
-                localSegmentsPostRefresh,
-                segmentInfosSnapshot,
-                storeDirectory,
-                translogFileGeneration,
-                replicationCheckpoint,
-                indexShard.getNodeId()
-            );
-        }
-    }
-
-    private void uploadNewSegments(
-        Collection<String> localSegmentsPostRefresh,
-        Map<String, Long> localSegmentsSizeMap,
-        ActionListener<Void> listener
-    ) {
-        Collection<String> filteredFiles = localSegmentsPostRefresh.stream().filter(file -> !skipUpload(file)).collect(Collectors.toList());
-        if (filteredFiles.size() == 0) {
-            logger.debug("No new segments to upload in uploadNewSegments");
-            listener.onResponse(null);
-            return;
-        }
-
-        logger.debug("Effective new segments files to upload {}", filteredFiles);
-        ActionListener<Collection<Void>> mappedListener = ActionListener.map(listener, resp -> null);
-        GroupedActionListener<Void> batchUploadListener = new GroupedActionListener<>(mappedListener, filteredFiles.size());
-        Directory directory = ((FilterDirectory) (((FilterDirectory) storeDirectory).getDelegate())).getDelegate();
-
-        for (String src : filteredFiles) {
-            // Initializing listener here to ensure that the stats increment operations are thread-safe
-            UploadListener statsListener = createUploadListener(localSegmentsSizeMap);
-            ActionListener<Void> aggregatedListener = ActionListener.wrap(resp -> {
-                statsListener.onSuccess(src);
-                batchUploadListener.onResponse(resp);
-                if (directory instanceof CompositeDirectory) {
-                    ((CompositeDirectory) directory).afterSyncToRemote(src);
-                }
-            }, ex -> {
-                logger.warn(() -> new ParameterizedMessage("Exception: [{}] while uploading segment files", ex), ex);
-                if (ex instanceof CorruptIndexException) {
-                    indexShard.failShard(ex.getMessage(), ex);
-                }
-                statsListener.onFailure(src);
-                batchUploadListener.onFailure(ex);
-            });
-            statsListener.beforeUpload(src);
-            remoteDirectory.copyFrom(storeDirectory, src, IOContext.DEFAULT, aggregatedListener, isLowPriorityUpload());
-        }
-    }
-
+    //todo
     boolean isLowPriorityUpload() {
         return isLocalOrSnapshotRecoveryOrSeeding();
     }
@@ -473,6 +264,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
      * @param file that needs to be uploaded.
      * @return true if the upload has to be skipped for the file.
      */
+    //todo: create a common class
     private boolean skipUpload(String file) {
         try {
             // Exclude files that are already uploaded and the exclude files to come up with the list of files to be uploaded.
@@ -497,55 +289,15 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
     }
 
     /**
-     * Updates the last refresh time and refresh seq no which is seen by remote store.
-     */
-    private void updateRemoteRefreshTimeAndSeqNo(long refreshTimeMs, long refreshClockTimeMs, long refreshSeqNo) {
-        segmentTracker.updateRemoteRefreshClockTimeMs(refreshClockTimeMs);
-        segmentTracker.updateRemoteRefreshTimeMs(refreshTimeMs);
-        segmentTracker.updateRemoteRefreshSeqNo(refreshSeqNo);
-    }
-
-    /**
      * Updates map of file name to size of the input segment files in the segment tracker. Uses {@code storeDirectory.fileLength(file)} to get the size.
      *
      * @param segmentFiles list of segment files that are part of the most recent local refresh.
      *
      * @return updated map of local segment files and filesize
      */
+    //todo: create a common class
     private Map<String, Long> updateLocalSizeMapAndTracker(Collection<String> segmentFiles) {
         return segmentTracker.updateLatestLocalFileNameLengthMap(segmentFiles, storeDirectory::fileLength);
-    }
-
-    private void updateFinalStatusInSegmentTracker(boolean uploadStatus, long bytesBeforeUpload, long startTimeInNS) {
-        if (uploadStatus) {
-            long bytesUploaded = segmentTracker.getUploadBytesSucceeded() - bytesBeforeUpload;
-            long timeTakenInMS = TimeValue.nsecToMSec(System.nanoTime() - startTimeInNS);
-            segmentTracker.incrementTotalUploadsSucceeded();
-            segmentTracker.updateUploadBytesMovingAverage(bytesUploaded);
-            segmentTracker.updateUploadBytesPerSecMovingAverage((bytesUploaded * 1_000L) / Math.max(1, timeTakenInMS));
-            segmentTracker.updateUploadTimeMovingAverage(timeTakenInMS);
-        } else {
-            segmentTracker.incrementTotalUploadsFailed();
-        }
-    }
-
-    /**
-     * On primary term update, we (re)initialise the remote segment directory to reflect the latest metadata file that
-     * has been uploaded to remote store successfully. This method also updates the segment tracker about the latest
-     * uploaded segment files onto remote store.
-     */
-    private void initializeRemoteDirectoryOnTermUpdate() throws IOException {
-        if (this.primaryTerm != indexShard.getOperationPrimaryTerm()) {
-            logger.trace("primaryTerm update from={} to={}", primaryTerm, indexShard.getOperationPrimaryTerm());
-            this.primaryTerm = indexShard.getOperationPrimaryTerm();
-            RemoteSegmentMetadata uploadedMetadata = this.remoteDirectory.init();
-
-            // During failover, the uploaded metadata would have names of files that have been uploaded to remote store.
-            // Here we update the tracker with latest remote uploaded files.
-            if (uploadedMetadata != null) {
-                segmentTracker.setLatestUploadedFiles(uploadedMetadata.getMetadata().keySet());
-            }
-        }
     }
 
     /**
@@ -553,6 +305,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
      * returned value of this method for scheduling retries in syncSegments method.
      * @return true iff the shard is a started with primary mode true or it is local or snapshot recovery.
      */
+    //todo: create a common class
     private boolean isReadyForUpload() {
         boolean isReady = indexShard.isStartedPrimary() || isLocalOrSnapshotRecoveryOrSeeding();
 
@@ -576,6 +329,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         return isReady;
     }
 
+    //todo: create a common class
     boolean isLocalOrSnapshotRecoveryOrSeeding() {
         // In this case when the primary mode is false, we need to upload segments to Remote Store
         // This is required in case of remote migration seeding/snapshots/shrink/ split/clone where we need to durable persist
@@ -585,39 +339,6 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
             && (indexShard.recoveryState().getRecoverySource().getType() == RecoverySource.Type.LOCAL_SHARDS
                 || indexShard.recoveryState().getRecoverySource().getType() == RecoverySource.Type.SNAPSHOT
                 || indexShard.shouldSeedRemoteStore());
-    }
-
-    /**
-     * Creates an {@link UploadListener} containing the stats population logic which would be triggered before and after segment upload events
-     *
-     * @param fileSizeMap updated map of current snapshot of local segments to their sizes
-     */
-    private UploadListener createUploadListener(Map<String, Long> fileSizeMap) {
-        return new UploadListener() {
-            private long uploadStartTime = 0;
-
-            @Override
-            public void beforeUpload(String file) {
-                // Start tracking the upload bytes started
-                segmentTracker.addUploadBytesStarted(fileSizeMap.get(file));
-                uploadStartTime = System.currentTimeMillis();
-            }
-
-            @Override
-            public void onSuccess(String file) {
-                // Track upload success
-                segmentTracker.addUploadBytesSucceeded(fileSizeMap.get(file));
-                segmentTracker.addToLatestUploadedFiles(file);
-                segmentTracker.addUploadTimeInMillis(Math.max(1, System.currentTimeMillis() - uploadStartTime));
-            }
-
-            @Override
-            public void onFailure(String file) {
-                // Track upload failure
-                segmentTracker.addUploadBytesFailed(fileSizeMap.get(file));
-                segmentTracker.addUploadTimeInMillis(Math.max(1, System.currentTimeMillis() - uploadStartTime));
-            }
-        };
     }
 
     /**
